Moves EnvironmentVariableReplacer static class into own file and hooks up SettingsReplacer to use it
