+++ src/Templater/SettingsReplacer.cs
@@ -1,5 +1,4 @@
-﻿using System.Collections;
-using System.Collections.Generic;
+﻿using System.Collections.Generic;
 using System.Linq;
 using System.Text.RegularExpressions;
 using log4net;
@@ -21,6 +20,8 @@ namespace Templater
 		{
 			globals = globals ?? new Environment();
 			UnusedKeys = new HashSet<string>();
+
+			input = EnvironmentVariableReplacer.SwapEnvironmentVariables(input);
 
 			foreach (var key in environment.Values.Keys.Where(key => !input.ContainsKey(key)).Where(key => !UnusedKeys.Contains(key.ToLower())))
 				UnusedKeys.Add(key.ToLower());
